feat: criado estrutura de valições e melhorado as models

+++ backend/src/controllers/UserController.js
@@ -5,7 +5,9 @@ class UserController {
     try {
       const newUser = request.body;
       if (!newUser.name || !newUser.email || !newUser.password) {
-        return reply.code(400).send({ message: 'Digite todos os dados' });
+        return reply
+          .code(400)
+          .send({ message: 'Campos obrigatórios ausentes' });
       }
 
       const user = await userService.createUser(newUser);
@@ -26,16 +28,17 @@ class UserController {
       const users = await userService.getUsers();
 
       if (users.length === 0) {
-        reply.code(404).send({ message: 'Nenhum usuário encontrado.' });
-      } else {
-        reply.code(200).send({ message: 'Usuários encontrados', users });
+        reply.code(404).send({ message: 'Registros não encontrados.' });
       }
 
-      reply.code(200).send(users);
+      reply.code(200).send({ message: 'Usuários encontrados', users });
     } catch (error) {
       console.error(error);
 
-      reply.code(500).send({ error: 'Usuários não encontrados' });
+      const status = error.status || 500;
+      const message = error.message || 'Erro interno no servidor.';
+
+      reply.code(status).send({ message });
     }
   }
 
@@ -45,10 +48,18 @@ class UserController {
 
       const user = await userService.getUserByID(id);
 
+      if (!user) {
+        reply.code(404).send({ message: 'Registro não encontrado.' });
+      }
+
       reply.code(200).send({ message: 'Usuário encontrado', user });
     } catch (error) {
       console.error(error);
-      reply.code(500).send({ error: '' });
+
+      const status = error.status || 500;
+      const message = error.message || 'Erro interno no servidor.';
+
+      reply.code(status).send({ message });
     }
   }
 
@@ -73,8 +84,8 @@ class UserController {
       }
 
       if (
-        'nome' in novosDados &&
-        (!novosDados.nome || novosDados.nome.trim() === '')
+        'name' in novosDados &&
+        (!novosDados.name || novosDados.name.trim() === '')
       ) {
         return reply.code(400).send({
           message: 'O campo "nome" não pode estar vazio.',
@@ -99,6 +110,12 @@ class UserController {
       const { id } = request.params;
       const { status } = request.body;
 
+      if (typeof status !== 'boolean' || status === undefined) {
+        return reply.code(400).send({
+          message: 'Status não enviado para atualização.',
+        });
+      }
+
       const user = await userService.toggleUserStatus(id, status);
 
       reply.code(200).send({
@@ -107,7 +124,11 @@ class UserController {
       });
     } catch (error) {
       console.error(error);
-      reply.code(500).send({ error: 'Usuário não desativado' });
+
+      const status = error.status || 500;
+      const message = error.message || 'Erro interno no servidor.';
+
+      reply.code(status).send({ message });
     }
   }
 
@@ -115,14 +136,46 @@ class UserController {
     try {
       const { id } = request.params;
 
+      if(!id) {
+        return reply.code(400).send({
+          message: 'Usuário não informado.',
+        });
+      }
+
       const user = await userService.deleteUser(id);
 
-      reply.code(204).send({ message: 'Usuário deletado', user });
+      console.log(user);
+
+      if (!user) {
+        reply.code(400).send({ message: 'Usuário não encontrado' });
+      }
+
+      reply.code(204).send({ message: 'Usuário deletado' });
     } catch (error) {
       console.error(error);
-      reply.code(500).send({ error: 'Usuário não deletado' });
+
+      const status = error.status || 500;
+      const message = error.message || 'Erro interno no servidor.';
+
+      reply.code(status).send({ message });
     }
   }
 }
 
 export default new UserController();
+
+// Situação	Código HTTP	Mensagem
+// Requisição com dados ausentes	400 Bad Request	Campos obrigatórios ausentes.
+// Registro duplicado	409 Conflict	Já existe um registro com este valor.
+// Registro não encontrado	404 Not Found	Registro não encontrado.
+// Usuário sem autenticação	401 Unauthorized	Autenticação requerida.
+// Usuário sem permissão	403 Forbidden	Permissão negada.
+// Erro do servidor	500 Internal Server Error	Ocorreu um erro no servidor.
+
+// Código do Prisma	Descrição	HTTP Status
+// P2002	Conflito de chave única (e.g., email duplicado).	409 Conflict
+// P2025	Registro não encontrado (ex.: update ou delete em ID inexistente).	404 Not Found
+// P2003	Violação de restrição de chave estrangeira.	400 Bad Request
+// P2004	Restrições de check falharam (valores inválidos no esquema).	422 Unprocessable Entity
+// P2014	Violação de restrição de relação no banco de dados.	400 Bad Request
+// P2016	Requisição inválida para um modelo que não existe no Prisma schema.	500 Internal Server Error
